change default stroke_linecap styling for all shapes
so that it matches the style used when drawing lines. remove opacity
setting for lines being drawn.

disable scroll events to prevent zooming when doing fast swiping in
surface pro 3 (windows 8.1)

in addition, reduce default expansion ratio to 1.0 from 3 (the
width/height of the drawing area was 3 times the width/height of the svg
drawing)